@@ -9,33 +9,35 @@
   },
   "private": true,
   "devDependencies": {
-    "@nestjs/schematics": "^8.0.0",
-    "@nestjs/testing": "^8.0.0",
-    "@nrwl/cli": "13.8.5",
-    "@nrwl/eslint-plugin-nx": "13.8.5",
-    "@nrwl/jest": "13.8.5",
-    "@nrwl/js": "13.8.5",
-    "@nrwl/linter": "13.8.5",
-    "@nrwl/nest": "13.8.5",
-    "@nrwl/node": "13.8.5",
-    "@nrwl/tao": "13.8.5",
-    "@nrwl/workspace": "13.8.5",
-    "@types/jest": "27.0.2",
-    "@types/node": "14.14.33",
-    "@typescript-eslint/eslint-plugin": "5.10.2",
-    "@typescript-eslint/parser": "5.10.2",
-    "eslint": "8.7.0",
+    "@nestjs/schematics": "9.0.4",
+    "@nestjs/testing": "9.3.9",
+    "@nrwl/cli": "15.8.5",
+    "@nrwl/eslint-plugin-nx": "15.8.5",
+    "@nrwl/jest": "15.8.5",
+    "@nrwl/js": "15.8.5",
+    "@nrwl/linter": "15.8.5",
+    "@nrwl/nest": "15.8.5",
+    "@nrwl/node": "15.8.5",
+    "@nrwl/workspace": "15.8.5",
+    "@types/jest": "29.4.0",
+    "@types/node": "18.7.1",
+    "@typescript-eslint/eslint-plugin": "5.54.0",
+    "@typescript-eslint/parser": "5.54.0",
+    "eslint": "8.15.0",
     "eslint-config-prettier": "^8.1.0",
     "eslint-plugin-prettier": "^3.4.0",
-    "jest": "27.2.3",
-    "prettier": "2.5.1",
-    "ts-jest": "27.0.5",
-    "typescript": "~4.3.5"
+    "jest": "29.4.3",
+    "jest-environment-jsdom": "28.1.3",
+    "prettier": "2.8.4",
+    "ts-jest": "29.0.5",
+    "ts-node": "10.9.1",
+    "typescript": "4.9.5",
+    "nx": "15.8.5"
   },
   "dependencies": {
-    "@nestjs/common": "^8.0.0",
-    "@nestjs/core": "^9.0.5",
-    "@nestjs/platform-express": "^8.0.0",
+    "@nestjs/common": "9.3.9",
+    "@nestjs/core": "9.3.9",
+    "@nestjs/platform-express": "9.3.9",
     "chalk": "4.1.2",
     "knex": "^2.4.0",
     "mysql2": "^2.3.0",
